Add initial DTOs and configuration for attendance service

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,11 +1,11 @@
 CREATE TABLE users (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     username VARCHAR(50) UNIQUE NOT NULL,
     role VARCHAR(20) NOT NULL
 );
 
 CREATE TABLE class_sessions (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     title VARCHAR(100) NOT NULL,
     start_time TIMESTAMP NOT NULL,
     end_time TIMESTAMP,
@@ -13,17 +13,17 @@ CREATE TABLE class_sessions (
 );
 
 CREATE TABLE attendance_records (
-    id BIGSERIAL PRIMARY KEY,
-    session_id BIGINT REFERENCES class_sessions(id),
-    user_id BIGINT REFERENCES users(id),
+    id SERIAL PRIMARY KEY,
+    session_id INT REFERENCES class_sessions(id),
+    user_id INT REFERENCES users(id),
     joined_at TIMESTAMP NOT NULL,
     left_at TIMESTAMP
 );
 
 CREATE TABLE session_events (
-    id BIGSERIAL PRIMARY KEY,
-    session_id BIGINT REFERENCES class_sessions(id),
-    user_id BIGINT REFERENCES users(id),
+    id SERIAL PRIMARY KEY,
+    session_id INT REFERENCES class_sessions(id),
+    user_id INT REFERENCES users(id),
     event_type VARCHAR(50) NOT NULL,
     payload TEXT,
     timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP
